Add categories options support
